Fix security issue regarding fan control
Until now, the dell-smm-hwmon kernel module was loaded with
restricted=1. which allowed non-root users to set the fan speed.
This however is a security risk, since a malicious program could
damage the hardware by disabling the fans.
Removing this option prevents such a scenario.
Also remove the possiblity of sourcing a user-provided config file,
since the script now always runs as root and sourcing arbitrary
files is unsafe.