Rewrite i8kctl in tcl
Rewrite i8kctl in tcl to simplify the project. Remove support
for the deprecated procfs interface since tcl scripts cannot
issue ioctl calls. Also remove support for reading fan speed
sensors and change the cli interface to better align with the
role of i8kctl as a program used to read/write fan states.

Nowadays the sensors program is the standard for displaying
sensor information, the main focus of i8kctl is to provide
functionality which is not provided by sensors.

Signed-off-by: Armin Wolf <W_Armin@gmx.de>