Add preliminary support for Tcl 9.0
We are already compatible with Tcl 9.0, the only thing blocking
us from running with a Tcl 9.0 interpreter is the fact that we
still check for compatibility with Tcl 8.6.

Drop this check to allow i8kutils to run with a Tcl 9.0 interpreter.
As soon as Tcl 9.0 is available in Debian we will fully migrate to
Tcl 9.0.

Signed-off-by: Armin Wolf <W_Armin@gmx.de>